refactor: move token
