+++ tiny/src/tokenizer.rs
@@ -1,21 +1,5 @@
+use crate::value_object::token::Token;
 use tailcall::tailcall;
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum Token {
-    KeywordColon,
-    KeywordPlus,
-    KeywordQuestion,
-    LiteralInt(i32),
-}
-
-pub fn token_to_string(t: Token) -> String {
-    match t {
-        Token::KeywordColon => String::from(":"),
-        Token::KeywordPlus => String::from("+"),
-        Token::KeywordQuestion => String::from("?"),
-        Token::LiteralInt(i) => format!("{}", i),
-    }
-}
 
 pub enum TokenizeError {
     ParseIntError,
