+++ tiny/src/value_object.rs
@@ -1,2 +1,3 @@
 pub mod ast;
 pub mod opcode;
+pub mod token;
